:mortar_board: Created test harness using GoogleTest
See /test/README.md

Currently there are 3 tests:
 * UtilsTestSuite / Func_SanitizeUtf8_ReplaceXff: Classic unit test, it feeds invalid UTF-8 character into `RoR::Utils::SanitizeUtf8String()` and expects the result to contain valid replacement.
 * ActorTestSuite / Method_ResetPosition_TranslateOnly: Another unit test, it invokes `Actor::ResetPosition(Ogre::Vector3, bool)` and checks the resulting node positions.
 * SimControllerTestSuite / Feat_QueueActorSpawn_Terrn2ActorRotated: A "feature test": Simulates spawning actor via "truck" keyword in ".tobj" file (with non-zero rotation) and checks resulting node positions. Motivated by bug #2334

+++ source/main/gui/GUIInputManager.h
@@ -39,7 +39,7 @@ public:
 
     void setMousePosition(int _x, int _y);
 
-    float getLastMouseMoveTime() { return m_last_mousemove_time->getMilliseconds(); };
+    float getLastMouseMoveTime() { return static_cast<float>( m_last_mousemove_time->getMilliseconds() ); };
 
     void SupressCursor(bool do_supress);
